Rename package, fix submission being pre-empted on shutdown
